fixed majority element
